pm-analyst/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.27.0",
    "google-adk>=1.25.1",
    "google-genai>=1.0.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",
    "msal>=1.28.0",
    "httpx>=0.27.0",
    "python-docx>=1.1.0",
]
